requirements.txt
numpy
pandas
scipy
tqdm
spacy
beautifulsoup4
tensorflow-gpu
pyyaml